Added layer check info and moved around a few things
