+++ packages/dds/shared-object-base/package.json
@@ -34,6 +34,7 @@
     "@fluidframework/core-interfaces": "^0.31.0",
     "@fluidframework/datastore": "^0.31.0",
     "@fluidframework/datastore-definitions": "^0.31.0",
+    "@fluidframework/gc-utils": "^0.31.0",
     "@fluidframework/protocol-definitions": "^0.1015.0",
     "@fluidframework/runtime-definitions": "^0.31.0",
     "@fluidframework/runtime-utils": "^0.31.0",
