+++ packages/dds/shared-object-base/src/sharedObject.ts
@@ -14,8 +14,9 @@ import {
     IChannelStorageService,
     IChannelServices,
 } from "@fluidframework/datastore-definitions";
+import { IGraphNode } from "@fluidframework/gc-utils";
 import { ISequencedDocumentMessage, ITree } from "@fluidframework/protocol-definitions";
-import { IChannelSummarizeResult, IGraphNode } from "@fluidframework/runtime-definitions";
+import { IChannelSummarizeResult } from "@fluidframework/runtime-definitions";
 import { convertToSummaryTreeWithStats, FluidSerializer } from "@fluidframework/runtime-utils";
 import { ChildLogger, EventEmitterWithErrorHandling } from "@fluidframework/telemetry-utils";
 import { SharedObjectHandle } from "./handle";
