+++ packages/runtime/container-runtime/src/dataStores.ts
@@ -19,7 +19,6 @@ import {
     IEnvelope,
     IFluidDataStoreChannel,
     IFluidDataStoreContextDetached,
-    IGraphNode,
     IInboundSignalMessage,
     InboundAttachMessage,
     ISummarizeResult,
@@ -37,7 +36,7 @@ import { BlobCacheStorageService, buildSnapshotTree, readAndParseFromBlobs } fro
 import { assert, Lazy } from "@fluidframework/common-utils";
 import { v4 as uuid } from "uuid";
 import { TreeTreeEntry } from "@fluidframework/protocol-base";
-import { normalizeAndPrefixGCNodeIds } from "@fluidframework/gc-utils";
+import { IGraphNode, normalizeAndPrefixGCNodeIds } from "@fluidframework/gc-utils";
 import { DataStoreContexts } from "./dataStoreContexts";
 import { ContainerRuntime, nonDataStorePaths } from "./containerRuntime";
 import {
