+++ packages/runtime/runtime-definitions/package.json
@@ -31,6 +31,7 @@
     "@fluidframework/container-definitions": "^0.31.0",
     "@fluidframework/core-interfaces": "^0.31.0",
     "@fluidframework/driver-definitions": "^0.31.0",
+    "@fluidframework/gc-utils": "^0.31.0",
     "@fluidframework/protocol-definitions": "^0.1015.0",
     "@types/node": "^10.17.24"
   },
