+++ packages/utils/gc-utils/src/test/garbageCollector.spec.ts
@@ -5,8 +5,8 @@
 
 import { strict as assert } from "assert";
 import { ITelemetryErrorEvent, ITelemetryLogger } from "@fluidframework/common-definitions";
-import { IGraphNode } from "@fluidframework/runtime-definitions";
-import { IGCResult, runGarbageCollection } from "../garbageCollector";
+import { runGarbageCollection } from "../garbageCollector";
+import { IGCResult, IGraphNode } from "../interfaces";
 
 describe("Garbage Collector", () => {
     let logger: ITelemetryLogger;
