+++ tools/build-tools/data/layerInfo.json
@@ -113,6 +113,14 @@
                     "Telemetry-Utils",
                     "Container-Definitions"
                 ]
+            },
+            "GC-Utils": {
+                "packages": [
+                    "@fluidframework/gc-utils"
+                ],
+                "deps": [
+                    "Base-Definitions"
+                ]
             }
         }
     },
@@ -159,7 +167,8 @@
                 "deps": [
                     "Container-Definitions",
                     "Driver-Utils",
-                    "Container-Utils"
+                    "Container-Utils",
+                    "GC-Utils"
                 ]
             },
             "Framework": {
